refactor(collection-plugin): extract CollectionToolbar.vue from CollectionView.vue (#2528)
Pull the search toolbar + flag-filter menu + view toggle out of
CollectionView.vue into a child CollectionToolbar.vue. Rendered DOM is
byte-equivalent (same root wrapper, same classes, every data-testid
unchanged incl. the dynamic collections-flag-chip-<key> /
collection-view-custom-<id> passed through as props/v-for).

The two click-outside menus (filter + "+" add-view) move fully into the
child, so each useClickOutside menuRef binds the wrapper it renders; the
child also closes both on collection switch (was the parent's slug watch).
searchQuery and flagFilters stay in the parent (they feed filteredItems /
tableFilteredItems / the persist watch) and are driven via v-model. The
own-property-safe chip-mode read is extracted to a shared, unit-tested
flagFilterModeOf() so parent and child share one Object.hasOwn guard.

Both menus' outside-click dismissal proven by e2e: the "+" menu by
collection-add-view-menu.spec.ts, the filter menu by a new assertion in
collection-flag-filter.spec.ts.
