+++ docs/shared-utils.md
@@ -114,7 +114,7 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | Path                                 | Helper                                                          | When to use                                                                                                                                                                                                                                                            |
 | ------------------------------------ | --------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `@mulmoclaude/core/collection` (`core/deriveAll.ts`) | `deriveAll(schema, record, refRecords)` / `resolveRowRefs(...)` | Evaluate every `derived` field on a collection record (saturating, cycle-safe, ref-deref via a preloaded target cache). The ONE derive implementation — used by the client rendering layer and the server's `manageCollection` enrichment; never reimplement the loop. |
-| `@mulmoclaude/collection-plugin/vue` (`collectionViewMode.ts`) | `applicableViewModes(schema)` / `customViewKey(id)` | The view modes a schema can render (table always; calendar/kanban gated on date/enum fields; `custom:<id>` per declared view), and the `custom:<id>` selector key. Used by `CollectionView`'s toggle and the dashboard's per-tile view picker — don't re-derive the gating. |
+| `@mulmoclaude/collection-plugin/vue` (`collectionViewMode.ts`) | `applicableViewModes(schema)` / `customViewKey(id)` / `flagFilterModeOf(filters, key)` | The view modes a schema can render (table always; calendar/kanban gated on date/enum fields; `custom:<id>` per declared view), the `custom:<id>` selector key, and the own-property-safe read of a flag chip's tri-state mode (shadowing-key safe, like core's `ownProp` but on the plugin side). Used by `CollectionView`'s toggle/table filter and `CollectionToolbar` — don't re-derive the gating or re-hand-roll the `Object.hasOwn` guard. |
 | `packages/core/src/collection/core/dynamicIcon.ts` (`@mulmoclaude/core/collection`) | `selectDynamicRecord(records, source, orderBy, recordsById?)` / `resolveIcon(record, spec, staticIcon, recordsById?)` / `firstDateField(schema)` | Pure resolver for `CollectionSchema.dynamicIcon` (data-driven launcher icon): pick the deciding source record, then map it through the first-match-wins `rules`. `recordsById` (source records keyed by primaryKey, default `{}`) resolves any `valueFrom` value reference inside a `where`. Server wraps it in `packages/core/src/collection/server/dynamicIcon.ts`'s `computeCollectionIcon(collection, opts?)` (fails soft to `fallback ?? schema.icon`). |
 | `packages/core/src/collection/core/ownProp.ts` | `ownProp(obj, key)` | Own-property read (`Object.hasOwn` gated) for a plain object indexed by a user/LLM/feed-controlled key. A bare `obj[key]` reaches inherited `Object.prototype` members, so a dangling ref/embed id like `"constructor"`/`"__proto__"` would resolve to the `Object` function instead of being absent — breaking the "missing target ⇒ null (em-dash)" fail-soft and dropping the key from the JSON response (#2322). Used by `deriveAll`'s `resolveRowRefs` and `server/derive.ts`'s embed projection. (`where.ts` keeps its own private copy for the dynamicIcon `where` evaluator — #2443.) |
 | `packages/core/src/collection/core/where.ts` (`@mulmoclaude/core/collection`) | `matchesWhere(where, record, recordsById?)` | SQL-like AND-of-conditions predicate (`eq`/`ne`/`in`/`gt`/`gte`/`lt`/`lte`/`contains`) used by `dynamicIcon`'s `source.where` / `rules[].where`. A condition's value is a literal `value` or a `valueFrom: { record, field }` reference resolved against `recordsById` (default `{}`); an unresolved reference makes the condition false for every op. Distinct from `whenMatches`/`CollectionWhen` (single-field, used by fields/actions) — don't conflate the two. |
